+++ EventListener/MinimalApiVersionRequestListener.php
@@ -37,7 +37,7 @@ class MinimalApiVersionRequestListener
      */
     public function onKernelRequest(RequestEvent $event): void
     {
-        if (!$event->isMasterRequest()) {
+        if (!$event->isMainRequest()) {
             return;
         }